feat: transfer sns token test script (#1132)
