+++ scripts/deploy.snses.mjs
@@ -1,15 +1,11 @@
 #!/usr/bin/env node
 
-import { HttpAgent } from '@dfinity/agent';
 import { IDL } from '@dfinity/candid';
 import { ICManagementCanister, InstallMode } from '@dfinity/ic-management';
-import { Ed25519KeyIdentity } from '@dfinity/identity';
-import { Secp256k1KeyIdentity } from '@dfinity/identity-secp256k1';
 import { Principal } from '@dfinity/principal';
-import { jsonReviver } from '@dfinity/utils';
 import { copyFile, readFile } from 'node:fs/promises';
 import { join } from 'node:path';
-import pemfile from 'pem-file';
+import { SNSES, loadLocalIdentity, localAgent } from './utils.mjs';
 
 await copyFile(
 	'./node_modules/@dfinity/ledger-icrc/dist/candid/icrc_ledger.idl.js',
@@ -48,9 +44,6 @@ const MINTER_ID = nextArg({ args, option: '--minterId' });
 
 const LEDGER_WASM_PATH = join(process.cwd(), 'target', 'ic', 'ckbtc_ledger.wasm.gz');
 const INDEX_WASM_PATH = join(process.cwd(), 'target', 'ic', 'ckbtc_index.wasm.gz');
-
-const SNS_JSON_FILE = join(process.cwd(), 'src', 'frontend', 'src', 'env', 'tokens.sns.json');
-const SNSES = JSON.parse((await readFile(SNS_JSON_FILE)).toString(), jsonReviver);
 
 // We just install a subset for local development because dfx/local replica ultimately fails if too many canisters are installed in a row.
 // See: https://forum.dfinity.org/t/too-many-open-files-os-error-24-state-manager-src-lib-rs33/18217/12?u=peterparker
@@ -148,42 +141,9 @@ const installIndex = async ({ agent, ledgerCanisterId, indexCanisterId }) => {
 	});
 };
 
-/**
- * Load identity from local PEM file
- */
+const identity = await loadLocalIdentity(PEM_FILE);
 
-const loadIdentity = async (pemFile) => {
-	const rawKey = (await readFile(pemFile)).toString();
-
-	const buf = pemfile.decode(rawKey);
-
-	if (rawKey.includes('EC PRIVATE KEY')) {
-		if (buf.length !== 118) {
-			throw 'expecting byte length 118 but got ' + buf.length;
-		}
-		return Secp256k1KeyIdentity.fromSecretKey(buf.subarray(7, 39));
-	}
-
-	if (buf.length !== 85) {
-		throw 'expecting byte length 85 but got ' + buf.length;
-	}
-	return Ed25519KeyIdentity.fromSecretKey(buf.subarray(16, 48));
-};
-
-const identity = await loadIdentity(PEM_FILE);
-
-/**
- * Init an agent to communicate with local replica
- */
-
-const localAgent = async () => {
-	const agent = new HttpAgent({ identity, fetch, host: 'http://127.0.0.1:4943/' });
-	await agent.fetchRootKey();
-
-	return agent;
-};
-
-const agent = await localAgent();
+const agent = await localAgent({ identity });
 
 /**
  * Deploy Snses
